% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile setups.R
\name{chk_bridge}
\alias{chk_bridge}
\title{check for potential bridge connection}
\usage{
chk_bridge(map_df)
}
\arguments{
\item{line1}{line one}

\item{line2}{line two}
}
\description{
check for potential bridge connection
}
