% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet map draw.R
\name{getLeafletOptions}
\alias{getLeafletOptions}
\title{set up leaflet options}
\usage{
getLeafletOptions(minZoom, maxZoom, ...)
}
\arguments{
\item{minZoom}{minimum zoom}

\item{maxZoom}{maximum zoom}

\item{...}{additional parameters from leaflet}
}
\description{
set up leaflet options
}
