% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile setups.R
\name{map_coord_setup}
\alias{map_coord_setup}
\title{map tiles coordinates setup}
\usage{
map_coord_setup(map_df)
}
\arguments{
\item{map_df}{map basic data frame}
}
\description{
map tiles coordinates setup
}
