% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile setups.R
\name{tile_coord_setup}
\alias{tile_coord_setup}
\title{tile coordinates setup}
\usage{
tile_coord_setup(row, col)
}
\arguments{
\item{row}{tile row position}

\item{col}{tile column position}
}
\description{
tile coordinates setup
}
